cast to bigint

+++ pkg/storage/unified/sql/testdata/sqlite--resource_version_insert-single path.sql
@@ -7,6 +7,6 @@ INSERT INTO "resource_version"
     VALUES (
         '',
         '',
-        CAST((julianday('now') - 2440587.5) * 86400000000.0 AS INTEGER)
+        CAST((julianday('now') - 2440587.5) * 86400000000.0 AS BIGINT)
     )
 ;
